# perf: faster standalone compact